requirement.txt
langchain==0.2.16
langchain-core==0.2.43
langchain-community==0.2.16
langchain-openai==0.1.23
langchain-pinecone==0.1.3
langchain-text-splitters==0.2.4
langsmith>=0.1.112,<0.2.0

flask==2.3.3
sentence-transformers==2.6.1
pypdf==4.2.0
python-dotenv==1.0.1

-e .